TINY-7829: moved selection commands to a separate module
